refactor: 기능에 맞게 파일 이름, 컴포넌트 이름 수정
