+++ src/pages/search/index.ts
@@ -1,3 +1,3 @@
 export { default as SearchPage } from './SearchPage';
-export * from './ingredient';
 export * from './food';
+export * from './ingredient';
