+++ src/pages/search/ingredient/components/IngredientItem.tsx
@@ -1,27 +1,42 @@
-import { RatingButton, Typography } from '@/components/common';
+import { Typography } from '@/components/common';
 import styled from '@emotion/styled';
 
 type IngredientItemProps = {
+  title: string;
   number: number;
-  ingredient: {
-    name: string;
-    rating: 'GOOD' | 'OK' | 'CAUTION';
-    reason: string;
-  };
+  summary: string;
+  whyGood: string[];
 };
 
-const IngredientItem = ({ number, ingredient }: IngredientItemProps) => {
+const IngredientItem = ({
+  title,
+  number,
+  summary,
+  whyGood,
+}: IngredientItemProps) => {
   return (
     <Container>
       <Box>
-        <Typography
-          variant='subtitle'
-          weight='bold'
-          as='h3'
-        >{`${number}. ${ingredient.name}`}</Typography>
-        <RatingButton variant={ingredient.rating} />
+        <Circle>
+          <Typography variant='body2' weight='bold' color='white'>
+            {number}
+          </Typography>
+        </Circle>
+        <Typography variant='body1' weight='bold'>
+          {title}
+        </Typography>
       </Box>
-      <Typography variant='body3'>{ingredient.reason}</Typography>
+      <Typography variant='body3'>{summary}</Typography>
+      <FoodBenefit>
+        <Typography variant='body3' weight='bold'>
+          ✅ 이런 점이 좋아요:
+        </Typography>
+        {whyGood.map((benefit, index) => (
+          <Typography key={`${benefit}-${index}`} variant='body3'>
+            {`${index + 1}. ${benefit}`}
+          </Typography>
+        ))}
+      </FoodBenefit>
     </Container>
   );
 };
@@ -39,4 +54,21 @@ const Box = styled.div`
   display: flex;
   gap: ${({ theme }) => theme.spacing[2]};
   align-items: center;
+  margin-bottom: ${({ theme }) => theme.spacing[1]};
+`;
+
+const Circle = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 24px;
+  height: 24px;
+  border-radius: 50px;
+  background-color: ${({ theme }) => theme.colors.primary};
+`;
+
+const FoodBenefit = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: ${({ theme }) => theme.spacing[1]};
 `;
